tests/wpt/web-platform-tests/html/browsers/history/the-location-interface/location_hash.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_hash</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <iframe id="srcdoc-iframe"
       srcdoc="<div style='height: 200vh'></div><div id='test'></div>"></iframe>
    <script>
    test(function () {
      window.history.pushState(1, document.title, '#x=1');
      var hash = location.hash;

      assert_equals(hash, "#x=1", "hash");

    }, "location hash");

    var t = async_test("Setting location.hash on srcdoc iframe");
    addEventListener("load", t.step_func_done(function() {
      var frameWin = document.getElementById("srcdoc-iframe").contentWindow;
      assert_equals(frameWin.location.href, "about:srcdoc");
      assert_equals(frameWin.scrollY, 0, "Should not have scrolled yet");
      frameWin.location.hash = "test";
      assert_equals(frameWin.location.href, "about:srcdoc#test");
      assert_true(frameWin.scrollY > frameWin.innerHeight,
                  "Should have scrolled by more than one viewport height");
    }));
    </script>
  </body>
</html>
